Update loading data to paging source

+++ app/src/main/java/com/example/funnumberfacts/data/NumberFact.kt
@@ -1,6 +1,6 @@
 package com.example.funnumberfacts.data
 
 data class NumberFact(
-    val number: Int?,
+    val number: Long?,
     val fact: String
 )
